rest api

+++ spring_jpa_querydsl/src/main/java/com/muselab/sample/jpa/querydsl/controller/MemberController.java
@@ -5,10 +5,13 @@ import com.muselab.sample.jpa.querydsl.domain.Team;
 import com.muselab.sample.jpa.querydsl.service.EmployeeService;
 import com.muselab.sample.jpa.querydsl.service.TeamService;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
+import org.springframework.http.HttpHeaders;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.*;
+
+import java.util.List;
 
 @RestController
 public class MemberController {
@@ -19,21 +22,59 @@ public class MemberController {
     @Autowired
     private TeamService teamService;
 
-    @RequestMapping("/employee/{id}")
-    public Employee getEmployee(@PathVariable long id) {
+    @RequestMapping(value = "/employee", method = RequestMethod.GET)
+    public ResponseEntity<List<Employee>> getEmployee() {
+
+        List<Employee> employeeList = employeeService.findAll();
+        if (employeeList == null || employeeList.size() == 0) return new ResponseEntity<List<Employee>>(HttpStatus.NO_CONTENT);
+
+        return new ResponseEntity<List<Employee>>(employeeList,HttpStatus.OK);
+    }
+
+    @RequestMapping(value = "/employee/{id}", method = RequestMethod.GET)
+    public ResponseEntity<Employee> getEmployee(@PathVariable long id) {
         System.out.println("id=" + id);
-        return employeeService.findById(id);
+
+        Employee employee = employeeService.findById(id);
+        if (employee == null) return new ResponseEntity<Employee>(HttpStatus.NO_CONTENT);
+
+        return new ResponseEntity<Employee>(employee,HttpStatus.OK);
     }
 
-    @RequestMapping("/employee/insert")
-    public Employee saveEmployee(@RequestBody Employee employee) {
-        return employeeService.save(employee);
+    @RequestMapping(value = "/employee", method = RequestMethod.POST)
+    public ResponseEntity<Employee> saveEmployee(@RequestBody Employee employee) {
+
+        Employee employee1 = employeeService.findById(employee.getEmpId());
+        if(employee1 != null) return new ResponseEntity<Employee>(HttpStatus.CONFLICT);
+
+        employeeService.save(employee);
+
+        return new ResponseEntity<Employee>(employee, HttpStatus.CREATED);
     }
 
-    @RequestMapping("/employee/delete/{id}")
-    public String deleteEmployee(@PathVariable long id) {
+    @RequestMapping(value = "/employee", method = RequestMethod.PUT)
+    public ResponseEntity<Employee> updateEmployee(@RequestBody Employee employee) {
+
+        System.out.println("employee="+employee.toString());
+        Employee employee1 = employeeService.findById(employee.getEmpId());
+
+        System.out.println("employee1="+employee.toString());
+
+        if(employee1 == null) return new ResponseEntity<Employee>(HttpStatus.NOT_FOUND);
+
+        employee1.setName(employee.getName());
+        employeeService.save(employee1);
+        return new ResponseEntity<Employee>(HttpStatus.OK);
+    }
+
+    @RequestMapping(value = "/employee/{id}", method = RequestMethod.DELETE)
+    public ResponseEntity<Void> deleteEmployee(@PathVariable long id) {
+        Employee employee1 = employeeService.findById(id);
+
+        if(employee1 == null) return new ResponseEntity<Void>(HttpStatus.NOT_FOUND);
+
         employeeService.delete(id);
-        return "OK";
+        return new ResponseEntity<Void>(HttpStatus.NO_CONTENT);
     }
 
     @RequestMapping("/team/insert")
